Assignment
/* Implement a stack data structure, which follows a Last In First Out (LIFO) Policy using an array.
Suppose the name of the array is arr, the topmost element of the stack is pointed by a variable
called top and the size of the stack is denoted by max_size.
Perform the following operations:
1. Push(x): push an element x into the stack. Display an
appropriate message if the stack is full.
2. Pop(): pop the topmost element from the stack. Display an
appropriate message if the stack is empty.
3. Display(): Print all the elements of the stack from, starting
from the topmost element.
*/